+++ tests/PaymentMeanBrandListTest.php
@@ -3,8 +3,9 @@
 /**
  * Title: OmniKassa payment mean brand list test
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.1.0
  * @since 1.1.0
@@ -16,12 +17,12 @@ class Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMeanBrandListTest extends WP_Uni
 	function test_constructor() {
 		$list = new Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMeanBrandList( array(
 			Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::IDEAL,
-			Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::MINITIX,
+			Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::INCASSO,
 		) );
 
 		$string = (string) $list;
 
-		$this->assertEquals( 'IDEAL, MINITIX', $string );
+		$this->assertEquals( 'IDEAL, INCASSO', $string );
 	}
 
 	/**
@@ -30,10 +31,10 @@ class Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMeanBrandListTest extends WP_Uni
 	function test_add_payment_method() {
 		$list = new Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMeanBrandList();
 		$list->add_payment_mean_brand( Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::IDEAL );
-		$list->add_payment_mean_brand( Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::MINITIX );
+		$list->add_payment_mean_brand( Pronamic_WP_Pay_Gateways_OmniKassa_PaymentMethods::INCASSO );
 
 		$string = (string) $list;
 
-		$this->assertEquals( 'IDEAL, MINITIX', $string );
+		$this->assertEquals( 'IDEAL, INCASSO', $string );
 	}
 }